Grab bag of clang fixes for Windows code
These should hopefully be uncontroversial, so I'm sending them to Darin
because he has global OWNERS approval on src/*.  Let me know if you want
to split this up and farm it out to individual components.

R=darin@chromium.org
CC=thakis@chromium.org
BUG=82385

Fix some -Wdeprecated-writable-strings warnings for wide string
literals.
  net/dns/dns_config_service_win_unittest.cc
  net/http/http_auth_sspi_win.cc
  net/http/http_auth_sspi_win.h
  ui/shell_dialogs/select_file_dialog_win_unittest.cc

Use reinterpret_cast<> to convert from void * to a function pointer.
  chrome/browser/policy/policy_loader_win.cc
  content/browser/gamepad/gamepad_platform_data_fetcher_win.cc

Remove "public" from a typedef.  It's not clear that that really means.
  chrome/browser/enumerate_modules_model_unittest_win.cc

Escape an unescaped backslash in a wide string iteral to prevent it from
being interpreted as a unicode literal ("\uXXX").
  chrome/browser/storage_monitor/test_portable_device_watcher_win.cc

Use base::UTF8ToUTF16() over ATL::CA2W() and avoid an unnecessary
c_str().  Clang can't convert from CA2WEX<> to string16.
  chrome/browser/ui/views/external_tab_container_win.cc

Fix a -Wenum-compare warning when asserting that two enum values are
equal.
  sandbox/win/src/policy_params.h

Don't define va_copy for clang.  It provides one in <stdarg.h>.
  third_party/leveldatabase/port/port_chromium.h

Review URL: https://chromiumcodereview.appspot.com/15200005

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@200814 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 4a93d92ce04a56741819b84045b6f89154ac1116

+++ sandbox/win/src/policy_params.h
@@ -32,8 +32,10 @@ POLPARAMS_BEGIN(FileName)
   POLPARAM(BROKER)   // TRUE if called from the broker.
 POLPARAMS_END(FileName)
 
-COMPILE_ASSERT(OpenFile::NAME == FileName::NAME, to_simplify_fs_policies);
-COMPILE_ASSERT(OpenFile::BROKER == FileName::BROKER, to_simplify_fs_policies);
+COMPILE_ASSERT(OpenFile::NAME == static_cast<int>(FileName::NAME),
+               to_simplify_fs_policies);
+COMPILE_ASSERT(OpenFile::BROKER == static_cast<int>(FileName::BROKER),
+               to_simplify_fs_policies);
 
 // Policy parameter for name-based policies.
 POLPARAMS_BEGIN(NameBased)
